+++ PirateTARpe23/Models/Islands/IslandDetailsViewModel.cs
@@ -12,6 +12,6 @@
 
         public int LevelRequirement { get; set; }
 
-        public double XPReward { get; set; }
+        public int XPReward { get; set; }
     }
 }
